Add files via upload
Main change is the separation of map tile info into the Map class and a creation of a Texture file. Also, the walls have been re-worked to consist only of Rects to reduce the overhead from the Sprite class. 

A simple start screen was included and therefore lead to the inclusion of a simple Button class.